# backend/foundation/source_artifact_openapi.yaml
openapi: 3.0.3
info:
  title: ISO Smart Source Artifact API
  version: 1.0.0
  description: Runtime contract for the seven repository-local operations declared by the immutable source OpenAPI artifact.
security:
  - BearerAuth: []
paths:
  /v1/onboarding/status:
    get:
      operationId: getOnboardingStatus
      summary: Read onboarding and ISO 9000 Foundation Gate state
      responses:
        '200':
          description: Tenant-scoped onboarding status and applicable, answer-key-free learning paths.
          content:
            application/json:
              schema:
                type: object
                required: [foundation_gate]
                properties:
                  foundation_gate:
                    $ref: '#/components/schemas/FoundationGateStatus'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/onboarding/organizations:
    get:
      operationId: getOnboardingOrganizations
      summary: List canonical QMS organizations in the authenticated tenant
      description: This is an ISO Smart local projection. It does not infer equivalence with a legacy or AdminApps organization identifier.
      responses:
        '200':
          description: Tenant-scoped canonical organizations available for onboarding commands.
          content:
            application/json:
              schema:
                type: object
                required: [organizations]
                properties:
                  organizations:
                    type: array
                    items:
                      type: object
                      required: [id, display_name]
                      properties:
                        id: {type: string, format: uuid}
                        display_name: {type: string}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/onboarding/organizational-profile:
    post:
      operationId: postOrganizationalProfile
      summary: Persist the step-10 organizational profile with tenant-scoped provenance
      description: Implementation-defined transport for the source-defined profile fields. The canonical organization must be selected from the authenticated tenant; this operation neither calls AdminApps nor creates an organization.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationalProfileRequest'}
      responses:
        '201':
          description: The persisted workflow transition, or an idempotent replay of it.
          content: {application/json: {schema: {$ref: '#/components/schemas/OnboardingTransitionResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/onboarding/document-references:
    post:
      operationId: postOnboardingDocumentReferences
      summary: Ingest step-12 metadata and hashes without uploading content bytes
      description: Implementation-defined transport for the existing metadata-only ingestion slice. It cannot assert full document content, parsing, licensing, or DocumentVersion creation.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OnboardingEvidenceIngestionRequest'}
      responses:
        '201':
          description: Evidence references and the onboarding transition were atomically persisted, or replayed.
          content: {application/json: {schema: {$ref: '#/components/schemas/OnboardingTransitionResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/recommendations:
    get:
      operationId: getRecommendations
      summary: List tenant recommendations
      parameters:
        - name: clause
          in: query
          required: false
          schema: {type: string, maxLength: 80}
      responses:
        '200':
          description: Advisory recommendations; no result authorizes execution.
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items: {$ref: '#/components/schemas/Recommendation'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/recommendations/{id}/basis:
    get:
      operationId: getRecommendationBasis
      summary: Read frozen normative and evidence basis
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Tenant-scoped recommendation provenance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecommendationBasisResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/approvals/{id}/decision:
    post:
      operationId: postApprovalDecision
      summary: Record an authorized human decision for an AgentDecision requiring review
      parameters:
        - $ref: '#/components/parameters/UuidId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalDecisionRequest'}
      responses:
        '200':
          description: Append-only decision persisted with event, outbox, and audit records.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApprovalDecisionResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/events:
    post:
      operationId: postDomainEvent
      summary: Publish an allowlisted tenant domain event through the transactional outbox
      description: Requires a validated bearer claim with scope domain:events:publish. Payment, tenant provisioning, Foundation Gate completion, and normative publication events are internal/external-boundary events and cannot be injected through this operation.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DomainEventRequest'}
      responses:
        '202':
          description: Event accepted or an identical event replay recognized.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DomainEventResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/evidence:
    post:
      operationId: postEvidence
      summary: Record immutable evidence and provenance
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EvidenceCreateRequest'}
      responses:
        '201':
          description: Evidence created with event, outbox, audit, tenant, and trace references.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvidenceCreateResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/learning/iso9000/attempts:
    post:
      operationId: postFoundationAttempt
      summary: Evaluate and persist an ISO 9000:2026 Foundation Gate attempt
      description: Score and pass status are calculated server-side. Critical questions require a correct answer. Correct options and explanations are returned only after submission.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FoundationAttemptRequest'}
      responses:
        '201':
          description: Immutable attempt and mastery recorded; completion event/outbox emitted only on pass.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FoundationAttemptResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/Unavailable'}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    BadRequest:
      description: Request schema or parameter is invalid.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unauthorized:
      description: Authentication is required.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Forbidden:
      description: Authenticated identity, tenant, or role is not authorized.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    NotFound:
      description: Resource does not exist in the authenticated tenant.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Conflict:
      description: State, idempotency, or aggregate version conflict.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unprocessable:
      description: A semantically invalid request or learning attempt was supplied.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unavailable:
      description: Tenant authority or PostgreSQL runtime roles are unavailable.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
    OrganizationalProfileRequest:
      type: object
      required: [organization_id, event_id, role, expertise_level, size_range, sites_count, countries, sector, certification_status]
      properties:
        organization_id: {type: string, format: uuid}
        event_id: {type: string, format: uuid, description: Idempotency key for the transition.}
        role: {type: string, maxLength: 120}
        expertise_level: {type: string, maxLength: 120}
        size_range: {type: string, maxLength: 120}
        sites_count: {type: integer, minimum: 0}
        countries: {type: array, minItems: 1, items: {type: string, maxLength: 120}}
        sector: {type: string, maxLength: 255}
        certification_status: {type: string, maxLength: 120}
        employees_count: {type: integer, minimum: 0}
    OnboardingEvidenceIngestionRequest:
      type: object
      required: [organization_id, event_id, items]
      properties:
        organization_id: {type: string, format: uuid}
        event_id: {type: string, format: uuid}
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [event_id, source_type, source_uri, content_hash, captured_at]
            properties:
              event_id: {type: string, format: uuid}
              source_type: {type: string, enum: [strategy, process, kpi, audit, complaint, supplier, document]}
              source_uri: {type: string, maxLength: 2000}
              content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
              captured_at: {type: string, format: date-time}
              document_version_id: {type: string, format: uuid, nullable: true}
    OnboardingTransitionResponse:
      type: object
      required: [workflow_id, step, status, transition_id, event_id, replay]
      properties:
        workflow_id: {type: string, format: uuid}
        step: {type: string}
        status: {type: string, enum: [complete]}
        transition_id: {type: string, format: uuid}
        event_id: {type: string, format: uuid}
        replay: {type: boolean}
        content_bytes_read: {type: boolean, description: Always false for document references.}
...
    FoundationGateStatus:
      type: object
      required: [required, status, unlocks_quality_baseline, paths, latest_attempt]
      properties:
        required: {type: boolean, enum: [true]}
        status: {type: string, enum: [content_unavailable, not_started, retraining_required, passed]}
        unlocks_quality_baseline: {type: boolean}
        latest_attempt:
          allOf: [{$ref: '#/components/schemas/FoundationAttemptSummary'}]
          nullable: true
        completion_attempt_id: {type: string, format: uuid, nullable: true}
        concept_mastery:
          type: array
          items:
            type: object
            required: [concept_key]
            properties:
              concept_key: {type: string}
              score: {type: string, nullable: true}
              last_assessed_at: {type: string, format: date-time, nullable: true}
              retraining_due_at: {type: string, format: date-time, nullable: true}
        paths:
          type: array
          items: {$ref: '#/components/schemas/LearningPath'}
...
    FoundationAttemptRequest:
      type: object
      required: [learning_path_id, answers]
      properties:
        learning_path_id: {type: string, format: uuid}
        answers: {type: array, minItems: 1, maxItems: 200, items: {$ref: '#/components/schemas/FoundationAnswer'}}
...
    Recommendation:
      type: object
      required: [id, title, body, confidence, assumptions, status, intended_autonomy, created_at]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        body: {type: string}
        confidence: {type: string}
        assumptions: {type: array, items: {type: string}}
        impact: {type: string, nullable: true}
        status: {type: string, enum: [proposed]}
        intended_autonomy: {type: string, pattern: '^A[0-4]$'}
        created_at: {type: string, format: date-time}
    RecommendationBasisResponse:
      type: object
      required: [recommendation_id, basis]
      properties:
        recommendation_id: {type: string, format: uuid}
        basis:
          type: array
          items:
            type: object
            required: [id, standard, edition, requirement_code, requirement_text, rule_id, rule_version, evidence_id, evidence_hash, rationale, model, trace_id, created_at]
            properties:
              id: {type: string, format: uuid}
              standard: {type: string}
              edition: {type: string}
              requirement_code: {type: string}
              requirement_text: {type: string}
              rule_id: {type: string, format: uuid}
              rule_version: {type: string}
              evidence_id: {type: string, format: uuid}
              evidence_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
              evidence_revision: {type: integer}
              rationale: {type: string}
              model: {type: object}
              trace_id: {type: string, format: uuid}
              created_at: {type: string, format: date-time}
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision: {type: string, enum: [approve, reject, request_changes]}
        comments: {type: string, maxLength: 4000}
    ApprovalDecisionResponse:
      type: object
      required: [approval_id, decision, event_id, outbox_id, audit_id]
      properties:
        approval_id: {type: string, format: uuid}
        decision: {type: string, enum: [approve, reject, request_changes]}
        event_id: {type: string, format: uuid}
        outbox_id: {type: string, format: uuid}
        audit_id: {type: string, format: uuid}
...
    DomainEventResponse:
      type: object
      required: [event_id, duplicate]
      properties:
        event_id: {type: string, format: uuid}
        outbox_id: {type: string, format: uuid}
        audit_id: {type: string, format: uuid}
        duplicate: {type: boolean}
...